refactor: renaming files using camel case
