+++ src/components/atoms/radio-group.tsx
@@ -4,7 +4,7 @@ import * as React from "react"
 import * as RadioGroupPrimitive from "@radix-ui/react-radio-group"
 import { CircleIcon } from "@heroicons/react/24/solid"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/libs/utils"
 
 function RadioGroup({
   className,
