+++ src/components/features/home/molecules/InfiniteReviewsCarousel.tsx
@@ -3,7 +3,7 @@
 import { useEffect, useState } from "react"
 import { motion } from "framer-motion"
 import { ReviewCard } from "../atoms"
-import { Review } from "@/lib/userReviews"
+import { Review } from "@/libs/userReviews"
 
 interface InfiniteReviewsCarouselProps {
   reviews: Review[]
