+++ src/querys/ReactQueryProvider.tsx
@@ -2,7 +2,7 @@
 
 import React from "react";
 import { QueryClientProvider } from "@tanstack/react-query";
-import { queryClient } from "@/query/react-query-client";
+import { queryClient } from "@/querys/ReactQueryClient";
 
 export function ReactQueryProvider({ children }: { children: React.ReactNode }) {
     return <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>;
